refactor(quotes): share validator scaffolding across order types
