+++ app/services/quote_versions/validators/one_off/business_validator.rb
@@ -5,6 +5,7 @@ module QuoteVersions
     module OneOff
       class BusinessValidator < ::BaseValidator
         include Currencies
+        include CurrencyValidation
 
         def initialize(result, quote_version:, billing_items:, scope:)
           @quote_version = quote_version
@@ -29,16 +30,6 @@ module QuoteVersions
         private
 
         attr_reader :quote_version, :billing_items, :scope
-
-        def validate_currency
-          currency = quote_version.currency
-
-          if currency.blank?
-            add_error(field: :currency, error_code: "value_is_mandatory") if scope == :approve
-          elsif self.class.currency_list.exclude?(currency)
-            add_error(field: :currency, error_code: "invalid_currency")
-          end
-        end
 
         def validate_add_ons
           add_ons.each_with_index do |add_on, index|
