+++ app/services/quote_versions/validators/subscription_creation/business_validator.rb
@@ -5,6 +5,7 @@ module QuoteVersions
     module SubscriptionCreation
       class BusinessValidator < ::BaseValidator
         include Currencies
+        include CurrencyValidation
 
         def initialize(result, quote_version:, billing_items:, scope:)
           @quote_version = quote_version
@@ -32,16 +33,6 @@ module QuoteVersions
         private
 
         attr_reader :quote_version, :billing_items, :scope
-
-        def validate_currency
-          currency = quote_version.currency
-
-          if currency.blank?
-            add_error(field: :currency, error_code: "value_is_mandatory") if scope == :approve
-          elsif self.class.currency_list.exclude?(currency)
-            add_error(field: :currency, error_code: "invalid_currency")
-          end
-        end
 
         # NOTE: payment terms are not validated yet, the quote-level field lands with LAGO-1529
         def validate_dates
