+++ app/services/quote_versions/validators/subscription_creation_validator.rb
@@ -2,37 +2,15 @@
 
 module QuoteVersions
   module Validators
-    class SubscriptionCreationValidator < ::BaseValidator
-      def initialize(result, quote_version:, scope:)
-        @quote_version = quote_version
-        @scope = scope
-
-        super
-      end
-
-      def valid?
-        structural = SubscriptionCreation::StructuralValidator.new(result, billing_items:, scope:)
-        return false unless structural.valid?
-
-        SubscriptionCreation::BusinessValidator.new(result, quote_version:, billing_items:, scope:).valid?
-      end
-
+    class SubscriptionCreationValidator < BaseOrderTypeValidator
       private
 
-      attr_reader :quote_version, :scope
-
-      def billing_items
-        @billing_items ||= normalized_billing_items
+      def structural_validator_class
+        SubscriptionCreation::StructuralValidator
       end
 
-      def normalized_billing_items
-        items = quote_version.billing_items || {}
-
-        if items.is_a?(Hash)
-          items.deep_stringify_keys
-        else
-          items
-        end
+      def business_validator_class
+        SubscriptionCreation::BusinessValidator
       end
     end
   end
